More linked list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -9,7 +9,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *temp;
-	size_t count = 0;
+	unsigned int count = 0;
 
 	temp = h;
